+++ src/extensions/uncheckSiteNavigation/loc/en-us.js
@@ -1,5 +1,16 @@
 define([], function() {
   return {
-    "Title": "Uncheck Site Navigation"
+    "Title": "Uncheck Site Navigation",
+    "PanelTitle": "New list navigation default",
+    "PanelDescription": "When this is on, the \"Show in site navigation\" box starts unticked whenever someone creates a list or library on this site. Anyone can still tick it before creating.",
+    "ToggleLabel": "Untick \"Show in site navigation\" by default",
+    "ToggleOn": "On",
+    "ToggleOff": "Off",
+    "Save": "Save",
+    "Cancel": "Cancel",
+    "Change": "Change",
+    "ReadOnlyNotice": "Only site owners can change this setting.",
+    "StatusEnabled": "This site is set to untick \"Show in site navigation\" for new lists and libraries.",
+    "StatusDisabled": "This site uses the SharePoint standard: new lists and libraries are added to the site navigation."
   }
 });